+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @graphtrace/shared
 
+## 1.8.1
+
 ## 1.8.0
 
 ### Minor Changes