+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphtrace/storage
 
+## 1.8.1
+
+### Patch Changes
+
+- @graphtrace/shared@1.8.1
+
 ## 1.8.0
 
 ### Patch Changes